Move ZarrOpenActions to new sc.fiji.ome.zarr.open package
The util package previously held ZarrOpenActions, but giving it its own
'open' package leaves room for future open-coordination code that may
depend on both ui and util without the need for a util→ui dependency.
ScriptUtils.getTemplate() is bumped to public so the relocated test can
still reach it from outside the util package.
