+++ src/test/java/sc/fiji/ome/zarr/plugins/DnDHandlerPluginTest.java
@@ -42,10 +42,10 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.file.Path;
 
-import sc.fiji.ome.zarr.ui.DnDActionChooser;
-import sc.fiji.ome.zarr.settings.ZarrOpeningSettings;
-import sc.fiji.ome.zarr.util.ZarrOpenActions;
+import sc.fiji.ome.zarr.open.ZarrOpenActions;
 import sc.fiji.ome.zarr.settings.ZarrOpenBehavior;
+import sc.fiji.ome.zarr.settings.ZarrOpeningSettings;
+import sc.fiji.ome.zarr.ui.DnDActionChooser;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
 class DnDHandlerPluginTest
